backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Telegram Bot
aiogram==3.4.0

# Database
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Scheduling
apscheduler==3.10.4

# Security
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
